[misc] making LGTM even happier.

+++ src/core/tb2binconstr.cpp
@@ -24,7 +24,7 @@ BinaryConstraint::BinaryConstraint(WCSP* wcsp, EnumeratedVariable* xx, Enumerate
     supportY = vector<Value>(sizeY, x->getInf());
     trwsM = vector<Cost>(max(sizeX, sizeY), MIN_COST);
 
-    costs = vector<StoreCost>(sizeX * sizeY, StoreCost(MIN_COST));
+    costs = vector<StoreCost>((size_t)sizeX * (size_t)sizeY, StoreCost(MIN_COST));
 
     for (unsigned int a = 0; a < x->getDomainInitSize(); a++)
         for (unsigned int b = 0; b < y->getDomainInitSize(); b++)
